Django/django-rest-framework/api-level-1/requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
